10.22_1 18:39 miao
2014.10.22
 1.更多按钮修改。
 2.欢迎页定2秒。
 3.设备标识改用Open UDID+Keychain来保存。
 4.4个动作的名字修改。
 5.注册完之后图片详情页更新头像。
 6.刚注册完后摇一摇和捣捣乱逻辑不对。
 mark 发包
 7.注册完点头像崩溃。
 8.Open UDID的更正。

+++ Tools/MMProgressHUD/MMProgressHUD.m
@@ -32,7 +32,8 @@ NSString * const MMProgressHUDAnimationKeyDismissAnimation = @"dismiss";
 
 NSUInteger const MMProgressHUDConfirmationPulseCount = 8;//Keep this number even
 
-CGFloat const MMProgressHUDStandardDismissDelay = 0.75f;
+//CGFloat const MMProgressHUDStandardDismissDelay = 0.75f;
+float const  MMProgressHUDStandardDismissDelay = 0.75f;
 
 CGSize const MMProgressHUDDefaultImageSize = {37.f, 37.f};
